# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/downloads": {
      "get": {
        "operationId": "list_downloads",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDownloadsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "start_download",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartDownloadResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/google-fonts/{family}/fetch": {
      "post": {
        "operationId": "fetch_google_font",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "description": "Google Fonts family name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/google-fonts/{family}/{file}": {
      "get": {
        "operationId": "get_google_font_file",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "description": "Google Fonts family name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file",
            "in": "path",
            "description": "Font filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "font/ttf": {}
            }
          }
        }
      }
    },
...
    "/operations/{id}": {
      "delete": {
        "operationId": "cancel_operation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Operation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/pipelines": {
      "post": {
        "operationId": "start_pipeline",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartPipelineResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DownloadProgress": {
        "type": "object",
        "required": ["id", "filename", "downloaded", "status"],
        "properties": {
          "downloaded": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "filename": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DownloadStatus"
          },
          "total": {
            "type": ["integer", "null"],
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "DownloadStatus": {
        "oneOf": [
          {
            "type": "object",
            "required": ["status"],
            "properties": {
              "status": {
                "type": "string",
                "enum": ["started"]
              }
            }
          },
          {
            "type": "object",
            "required": ["status"],
            "properties": {
              "status": {
                "type": "string",
                "enum": ["downloading"]
              }
            }
          },
          {
            "type": "object",
            "required": ["status"],
            "properties": {
              "status": {
                "type": "string",
                "enum": ["completed"]
              }
            }
          },
          {
            "type": "object",
            "required": ["reason", "status"],
            "properties": {
              "reason": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": ["failed"]
              }
            }
          }
        ]
      },
...
      "ListDownloadsResponse": {
        "type": "object",
        "required": ["downloads"],
        "properties": {
          "downloads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DownloadProgress"
            }
          }
        }
      },
...
      "PageId": {
        "type": "string",
        "format": "uuid"
      },
...
      "Region": {
        "type": "object",
        "required": ["x", "y", "width", "height"],
        "properties": {
          "height": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "width": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "x": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "y": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "StartDownloadRequest": {
        "type": "object",
        "required": ["modelId"],
        "properties": {
          "modelId": {
            "type": "string",
            "description": "Package id, as declared via `declare_hf_model_package!`\n(e.g. `\"model:comic-text-detector:yolo-v5\"`)."
          }
        }
      },
      "StartDownloadResponse": {
        "type": "object",
        "required": ["operationId"],
        "properties": {
          "operationId": {
            "type": "string",
            "description": "Operation id. Reusing the package id keeps ids meaningful for clients\nwatching progress events."
          }
        }
      },
      "StartPipelineRequest": {
        "type": "object",
        "required": ["steps"],
        "properties": {
          "defaultFont": {
            "type": ["string", "null"]
          },
          "pages": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/PageId"
            },
            "description": "`None` → whole project, `Some(pages)` → just those pages."
          },
          "region": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Region",
                "description": "Optional bounding-box hint for inpainter engines (repair-brush)."
              }
            ]
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Engine ids (`inventory::submit!` ids) to run in order."
          },
          "systemPrompt": {
            "type": ["string", "null"]
          },
          "targetLanguage": {
            "type": ["string", "null"]
          }
        }
      },
      "StartPipelineResponse": {
        "type": "object",
        "required": ["operationId"],
        "properties": {
          "operationId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}